Added more unit tests.

+++ Token/Data/TokenUserData.cs
@@ -11,7 +11,6 @@ namespace Token.Data
   {
     private static readonly DateTime EPOCH_DATE = new DateTime(1970, 1, 1);
     private static readonly int TTL_MINUTES = Int32.Parse(Configuration.File.GetSection("Application")["DataTimeToLiveMinutes"]);
-
     private ILogger<TokenUserData> _logger;
     private ITokenUserRepository _tokenUserRepository;
 
@@ -23,11 +22,21 @@ namespace Token.Data
 
     public async Task Delete(string id)
     {
+      if (String.IsNullOrWhiteSpace(id))
+      {
+        throw new ArgumentNullException("id");
+      }
+
       await _tokenUserRepository.Delete(id);
     }
 
     public async Task<bool> Exists(string id)
     {
+      if (String.IsNullOrWhiteSpace(id))
+      {
+        throw new ArgumentNullException("id");
+      }
+
       TokenUser tokenUser = await _tokenUserRepository.Load(id);
 
       return tokenUser != null;
@@ -35,6 +44,11 @@ namespace Token.Data
 
     public async Task<TokenUser> Get(string id)
     {
+      if (String.IsNullOrWhiteSpace(id))
+      {
+        throw new ArgumentNullException("id");
+      }
+
       TokenUser tokenUser = await _tokenUserRepository.Load(id);
 
       return tokenUser;
@@ -42,6 +56,11 @@ namespace Token.Data
 
     public async Task Save(TokenUser tokenUser)
     {
+      if (tokenUser == null)
+      {
+        throw new ArgumentNullException("tokenUser");
+      }
+
       DateTime utcNow = DateTime.UtcNow;
 
       tokenUser.CreateDate = tokenUser.CreateDate ?? utcNow;
@@ -52,14 +71,14 @@ namespace Token.Data
       await _tokenUserRepository.Save(tokenUser);
     }
 
-    private static long GetTTL(DateTime? dateTime)
+    public static long GetTTL(DateTime? dateTime)
     {
-      if (dateTime.HasValue)
+      if (dateTime == null || !dateTime.HasValue)
       {
-        return (long)(dateTime.Value.AddMinutes(TTL_MINUTES) - EPOCH_DATE).TotalSeconds; ;
+        throw new ArgumentNullException("dateTime");
       }
 
-      throw new ArgumentNullException("dateTime");
+      return (long)(dateTime.Value.AddMinutes(TTL_MINUTES) - EPOCH_DATE).TotalSeconds;
     }
   }
 }
